Pull libraries from GitHub per default during build.sh Package using the available libraries use version.txt to get the artoolkitX dependencies with the correct version

+++ Source/packaging/package.sh
@@ -15,6 +15,7 @@ UNITY_VERSION=$1
 
 # Get our location.
 OURDIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+ARUNITYX_HOME=$OURDIR/../.. 
 
 #Check if we are using the submodule to build
 if [  -f "$OURDIR/../Extras/artoolkitX/LICENSE.txt" ] && [ -z $ARTOOLKITX_HOME ]; then
@@ -28,17 +29,27 @@ if [ ! -f ${CONFIG_LOCATION} ]; then
     #If only 'macos' was built config.h is in a different location
     CONFIG_LOCATION="${ARTOOLKITX_HOME}/SDK/Frameworks/ARX.framework/Headers/AR/config.h"
 fi
+
+# If there is still no config location then there probably is no clone of the artoolkitX source available
 if [ ! -f ${CONFIG_LOCATION} ]; then
-    #Print if we didn't built at all.
-    echo "You need to run ./build.sh <plattform> before packaging."
+    # Check if the plugins are available
+    if [ -f $ARUNITYX_HOME/Source/Package/Assets/Plugins/Android/arxjUnity.jar ] ||  [ -f $ARUNITYX_HOME/Source/Package/Assets/Plugins/iOS/libARX.a ] || [ -f $ARUNITYX_HOME/Source/Package/Assets/Plugins/x86_64/ARX.dll ] || [ -f $ARUNITYX_HOME/Source/Package/Assets/Plugins/ARX.bundle ]; then
+        # Plugins are available, build using the version number inside version.txt
+        VERSION=`cat $ARUNITYX_HOME/version.txt`
+    else
+        #Print if we didn't built at all.
+        echo "You need to run ./build.sh <plattform> before packaging."
+        exit -1
+    fi
+else
+    VERSION=`sed -En -e 's/.*AR_HEADER_VERSION_STRING[[:space:]]+"([0-9]+\.[0-9]+(\.[0-9]+)*)".*/\1/p' ${CONFIG_LOCATION}`
+    # If the tiny version number is 0, drop it.
+    VERSION=`echo -n "${VERSION}" | sed -E -e 's/([0-9]+\.[0-9]+)\.0/\1/'`
 fi
 
-VERSION=`sed -En -e 's/.*AR_HEADER_VERSION_STRING[[:space:]]+"([0-9]+\.[0-9]+(\.[0-9]+)*)".*/\1/p' ${CONFIG_LOCATION}`
-# If the tiny version number is 0, drop it.
-VERSION=`echo -n "${VERSION}" | sed -E -e 's/([0-9]+\.[0-9]+)\.0/\1/'`
+
 
 # Rename version, where appropriate.
-ARUNITYX_HOME=$OURDIR/../.. 
 sed -Ei "" "s/artoolkitX for Unity Version (([0-9]+\.[0-9]+)(\.[0-9]+)?(r[0-9]+)?)/artoolkitX for Unity Version $VERSION/" $ARUNITYX_HOME/Source/Package/Assets/ARToolKitX-Unity/Scripts/Editor/ARToolKitMenuEditor.cs
 
 # Build the unitypackage.
